requirements/base.txt
Pillow==6.1.0  # https://github.com/python-pillow/Pillow
redis==3.3.8  # https://github.com/antirez/redis


# Django
# ------------------------------------------------------------------------------
django==2.2.5  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.2.0  # https://github.com/jazzband/django-model-utils
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.3  # https://github.com/django-compressor/django-compressor
django-redis==4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.10.3  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
django-filter==2.2.0
requests==2.21.0
drf-yasg==1.16.1
mysqlclient==1.4.4
packaging==19.1
qiniu==7.2.6
django-mptt==0.10.0
elasticsearch-dsl==7.0.0